docker_worker/code/requirements.txt
Columnar==1.4.1
h5py==3.8.0
haversine==2.8.0
ipython==8.13.2
numpy==1.23.5
pandas==2.0.1
plotly==5.15.0
pvlib==0.9.5
pyephem==9.99
pymongo==4.4.1
scipy==1.11.1
sorcery==0.2.2
xarray==2023.5.0
polars==0.19.19
python-dotenv==1.0.1
